Use guint instead of ulong for signal ID.
2005-01-31  Hans Petter Jansson  <hpj@novell.com>

        * gui/e-cal-config.c (_ECalConfigPrivate): Use guint instead of ulong
        for signal ID.

svn path=/trunk/; revision=28640

+++ calendar/gui/e-cal-config.c
@@ -30,7 +30,7 @@ static GObjectClass *ecp_parent_class;
 static GObjectClass *ecph_parent_class;
 
 struct _ECalConfigPrivate {
-	ulong source_changed_id;
+	guint source_changed_id;
 };
 
 #define _PRIVATE(o) (g_type_instance_get_private ((GTypeInstance *)o, e_cal_config_get_type ()))
